renderer: use int for physical coordinates
now that all coordinates in the renderer
use physical coordinates we can return to
use integer for all coordinates.
this has the benefit that custom renderer
that use 2d blitting do not have to round
internally

+++ wlcs_anvil/src/renderer.rs
@@ -124,7 +124,7 @@ impl Frame for DummyFrame {
     type Error = SwapBuffersError;
     type TextureId = DummyTexture;
 
-    fn clear(&mut self, _color: [f32; 4], _damage: &[Rectangle<f64, Physical>]) -> Result<(), Self::Error> {
+    fn clear(&mut self, _color: [f32; 4], _damage: &[Rectangle<i32, Physical>]) -> Result<(), Self::Error> {
         Ok(())
     }
 
@@ -132,8 +132,8 @@ impl Frame for DummyFrame {
         &mut self,
         _texture: &Self::TextureId,
         _src: Rectangle<f64, Buffer>,
-        _dst: Rectangle<f64, Physical>,
-        _damage: &[Rectangle<f64, Physical>],
+        _dst: Rectangle<i32, Physical>,
+        _damage: &[Rectangle<i32, Physical>],
         _src_transform: Transform,
         _alpha: f32,
     ) -> Result<(), Self::Error> {
